+++ client/public/Styles/info.scss
