+++ client/public/Styles/navBar.scss
